improve comments
